Refactor: Extracts reusable code across test files
